requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-cors-middleware==1.4.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==1.0.0
ecdsa==0.13.2
future==0.17.1
idna==2.8
iso3166==1.0
jwt==0.6.1
oauthlib==3.1.0
pyasn1==0.4.6
pycparser==2.19
PyJWT==1.7.1
python-jose==3.0.1
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
